% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/case.R
\name{ignore.case}
\alias{ignore.case}
\alias{ignore.case.character}
\alias{ignore.case.default}
\alias{ignore.case.pattern}
\alias{ignore.case.searchable}
\alias{use.case}
\alias{use.case.character}
\alias{use.case.default}
\alias{use.case.pattern}
\alias{use.case.searchable}
\title{Turn on/off case sensitivity for searchable targets and patterns}
\usage{
ignore.case(object)

\method{ignore.case}{pattern}(object)

\method{ignore.case}{searchable}(object)

\method{ignore.case}{character}(object)

\method{ignore.case}{default}(object)

use.case(object)

\method{use.case}{pattern}(object)

\method{use.case}{searchable}(object)

\method{use.case}{character}(object)

\method{use.case}{default}(object)
}
\arguments{
\item{object}{search pattern or target

\code{ignore.case} and \code{use.case} control the case sensitivity of the
matching

The default is to preform case sensitive matching.}
}
\description{
Functions for affecting the case sensitivity of matching.
}
\examples{
"pattern" \%>\% use.case
  "pattern" \%>\% ignore.case
}
\seealso{
\code{\link{match.modifiers}} \cr
}

