% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/coll.R
\name{coll}
\alias{coll}
\alias{coll.SearchableOrPattern}
\alias{coll.character}
\alias{coll.default}
\title{coll}
\usage{
coll(object, ...)

\method{coll}{default}(object, ...)

\method{coll}{character}(object, ...)

\method{coll}{SearchableOrPattern}(object, ...)
}
\arguments{
\item{object}{to make specification}

\item{...}{additional arguments passes to \code{\link{pattern}}

\code{coll}}
}
\description{
Creates or modifies the search type to use coll matching
}
\examples{
pat <- coll("a")
  detect( c('alpha','beta'), pat )
}
\seealso{
\code{\link{pattern}}
}

