% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/detect.R
\name{detect}
\alias{detect}
\title{detect}
\usage{
detect(str, pattern)
}
\arguments{
\item{str}{seachable target}

\item{pattern}{method for searching}
}
\value{
logical; vector indicating which elements match
}
\description{
which elements matching the search pattern
}
\note{
- may not export this function
}
\seealso{
.matches
}

