% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fixed.R
\name{fixed}
\alias{fixed}
\alias{fixed.SearchableOrPattern}
\alias{fixed.character}
\alias{fixed.default}
\title{fixed}
\usage{
fixed(object, ...)

\method{fixed}{default}(object, ...)

\method{fixed}{character}(object, ...)

\method{fixed}{SearchableOrPattern}(object, ...)
}
\arguments{
\item{object}{to make specification}

\item{...}{additional arguments passes to \code{\link{pattern}}

\code{fixed}}
}
\description{
Creates or modifies the search type to use fixed matching
}
\examples{
pat <- fixed("a")
  detect( c('alpha','beta'), pat )
}
\seealso{
\code{\link{pattern}}
}

