% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/invert.R
\name{invert}
\alias{invert}
\title{Invert a structure by swapping keys and values}
\usage{
invert(x)
}
\arguments{
\item{x}{object to invert}
}
\value{
A character vector in which the names are the former valeus.
}
\description{
Invert a structure by swapping keys and values
}
\details{
Inverts named vectors
}
\note{
- currently applies to atomic vectors only
  - apply to list (recursive structures), data.frames, matrices and arrays
  - invert might be an ambiguous name ... call it swap_kv? kvswap? swapKV? swapNV?
}
\examples{
v <- 1:26
 names(v) <- letters

 invert(v)

 l <- as.list(v)
}

