% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.string.R
\name{is.string}
\alias{is.string}
\title{Test if an object is a string}
\usage{
is.string(x)
}
\arguments{
\item{x}{Equivalent to: \cr
\code{ is.character(x) && length(x) == 1 }}
}
\description{
A string is a one element character vector
}

