% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Class-pattern.R
\docType{class}
\name{pattern-class}
\alias{coll}
\alias{fixed}
\alias{pattern}
\alias{pattern-class}
\alias{regex}
\alias{show,pattern-method}
\alias{standard}
\title{Defines a search pattern}
\usage{
pattern(object = NULL, type = NULL, ...)

\S4method{show}{pattern}(object)

regex(object, ...)

coll(object, ...)

fixed(object, ...)

standard(object, ...)
}
\arguments{
\item{object}{character or pattern;}

\item{type}{character; the type of match: standard (default), regex, coll,
fixed.}

\item{....}{additional arguments to be passed to \code{stri_opts_*} functions.
       See details.


The \strong{pattern} class defines how the search is conducted.

The function \code{pattern} is the constructor for the class. It takes a
\code{'string'} can be used to define a pattern that controls matching
against a searchable target. Most often the user will want to use the type
specific functions: \code{regex}, \code{coll}, \code{fixed} or \code{basic}.
Eacb is described below.

These are closely related to the}
}
\description{
Defines how searches are conducted against a searchable target
}
\section{standard}{


The default is \code{standard} matching which performs matching
as base R would. This is equivalent to \code{fixed} and
\code{case_insensitive = FALSE}. Though the internal matching is sed.
}

\section{regex}{


\code{regex} matching takes a regular expression for matching using the
\code{stri_*_regex} functions.
}

\section{coll}{

...
}

\section{fixed}{

...
}
\examples{
pattern('hello')
  pattern('hello', type="regex", boundary="starts_with", )
}

