% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/regex.R
\name{regex}
\alias{regex}
\alias{regex.SearchableOrPattern}
\alias{regex.character}
\alias{regex.default}
\title{regex}
\usage{
regex(object, ...)

\method{regex}{default}(object, ...)

\method{regex}{character}(object, ...)

\method{regex}{SearchableOrPattern}(object, ...)
}
\arguments{
\item{object}{to make specification}

\item{...}{additional arguments passes to \code{\link{pattern}}

\code{regex}}
}
\description{
Creates or modifies the search type to use regular expression matching
}
\examples{
pat <- regex("a.+")
  detect( c('alpha','beta'), pat )
}
\seealso{
\code{\link{pattern}}
}

