% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/reverse.lookup.R
\name{reverse.lookup}
\alias{reverse.lookup}
\title{Perform a reverse lookup on searchables}
\usage{
reverse.lookup(string)
}
\arguments{
\item{string}{pattern for which to match against an objects values

\code{reverse.lookup} sets and toggles the logical attribute with name
\code{reverse.lookup}. Actual implementation of the reverse lookup is
performed in the Extract methods.


In order to perform a reverse lookup, values must be converted to character
names.}
}
\value{
a string object with the \code{reversse.lookup} attribute set.
}
\description{
This function causes the pattern search to be performed against an object's
values instead of its names
}
\note{
What happens if there are two reverse.lookups
}
\section{reverse.lookup}{


When performing a reverse lookup, values (not names) are searched. The
corresponding names are returned.  NOTE: this is highly experimental and only
works for atomic vectors. It is uncertain how this might be applied to
recursive structures like lists.
}
\examples{
reverse.lookup("string")
}
\seealso{
The \code{invert} function in the \code{hash} package \cr
}

