% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Class-SearchableOrPattern.R
\docType{class}
\name{SearchableOrPattern-class}
\alias{SearchableOrPattern-class}
\title{Class: SearchableOrPattern}
\description{
This is a class used internally to specify an object as either a searchable
or pattern. Both of these classes inherit from this class
}
\section{Slots}{

\describe{
\item{\code{.Data}}{Searchables; an object that can be made searchable, currently
a vector or list.}

\item{\code{type}}{character the}

\item{\code{options}}{list; list of stringi-type options

\code{SearchableOrPattern} wraps one of the classes that a \code{Searchables}
object.}
}}

