% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Class-SearchableOrPattern.R
\docType{class}
\name{SearchableOrPattern-class}
\alias{SearchableOrPattern}
\alias{SearchableOrPattern-class}
\title{Superclass of searchable and pattern}
\description{
Defines a superclas of classes \code{\link{searchable}} and
\code{\link{pattern}}
}
\section{Slots}{

\describe{
\item{\code{type}}{character; type of search performed; one of "std" (default),
"regex", "fixed", "coll", or "charclass". See details.}

\item{\code{options}}{list; name = value pairs for search options used.

The \code{SearchaleOrPattern} class defines the attributes common to both
searchables and patterns: \code{type} and \code{...}. Mainly this is just a
bookkeeping convention to simplify writing methods that change search
behavior, e.g. \code{regex}. This allows writing a single functions that
changes the search behavior at either the search target or search pattern or
\code{pattern}

The constructor function is not meant to be used directly and is not
exported.}
}}
\seealso{
\code{\link{searchable}} \cr
  \code{\link{pattern}}
}

