% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/std.R
\name{std}
\alias{std}
\alias{std.SearchableOrPattern}
\alias{std.character}
\alias{std.default}
\title{Use/revert to standard matching}
\usage{
std(object, ...)

\method{std}{default}(object, ...)

\method{std}{character}(object, ...)

\method{std}{SearchableOrPattern}(object, ...)
}
\arguments{
\item{object}{to make specification}

\item{...}{additional arguments passes to \code{\link{pattern}}

\code{std} causes match behavior to revert to R's default. It is a non-op
stringr-style match modifier that indicates that no special matching
behavior should be used.

The function mainly exist to revert clear previos modifiers.}
}
\description{
Creates or modifies the search type to use default R matching
}
\examples{
ptrn <- std("a")
  detect( c('alpha','beta'), ptrn )
}
\seealso{
\code{\link{pattern}}
}

