% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/case.R
\name{ignore.case}
\alias{case.insensitive}
\alias{case.insensitivity}
\alias{case.sensitive}
\alias{case.senstivity}
\alias{ignore.case}
\alias{use.case}
\title{Turn on/off case sensitivity for searching}
\usage{
ignore.case(object)

case.insensitive(object)

use.case(object)

case.sensitive(object)
}
\arguments{
\item{object}{search pattern or target

\code{ignore.case} and \code{case.insensitive} are synonyms.
\code{use.case} and \code{case.sensitive} are synonyms.

The default is to preform case sensitive matching.}
}
\description{
Functions for affecting the case sensitivity of matching.
}
\examples{
"pattern" \%>\% use.case
  "pattern" \%>\% ignore.case
}
\seealso{
\code{\link{match.modifiers}} \cr
}

