% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/match-modifiers.R
\name{perl}
\alias{exact}
\alias{fixed}
\alias{ignore.case}
\alias{partial}
\alias{perl}
\title{Match modifiers}
\usage{
perl(object)

ignore.case(object)

fixed(object)

partial(object)

exact(object)
}
\arguments{
\item{object}{string or target object to apply match modifiers to

Match modifiers control how a search pattern is matched against the search
target. They can be applied to either the pattern or target. Generally,
modifers applied to the target are applied when the object is created.
Modifiers applied to the search string take priority, are applied after and
override those of applied to the target.

Modifiers \code{ignore.case}, \code{perl} and \code{fixed} follow closely the
implementations by the \strong{stringr} package.}
}
\description{
Functions affecting how matching occurs. The modifiers can be applied to
either the search pattern or the search target.
}
\section{ignore.case}{


Match modifier \code{ignore.case} performs case insensitive matching against
the names of the target. \code{ignore.case} is incompatible with
\code{fixed}. Application of \code{ignore.case} overrides previous calls to
\code{fixed}.
}

\section{perl}{


Match modifier \code{perl} performs matches against the targets names using
PCREs. \code{perl} is incompatible with
\code{fixed}. Application of \code{perl} overrides previous calls to
\code{fixed}.
}

\section{fixed}{


Match modifier \code{fixed} performs a fixed string match instead of using a
regular expression. This can yield substantial speed ups, if regular
expression matching is not needed.  Fixed mathcing requires that the search
target \strong{contains} the search string. For exact matching of the search
string see \code{exact}.

\code{fixed} is incompatible with
\code{perl} and \code{ignore.case}.  Application of \code{fixed} overrides
previous calls to the others.
}

\section{partial}{


This is a synonym for \code{\link[stringr]{fixed}}
}

\section{exact}{


Match modifier \code{exact}, set matches to performs an exact match against
the target's object names. This is the default behavior of base. It is
incompatible with all other match modifiers. Setting \code{exact} sets the
other match modifiers to \code{FALSE}.
}
\examples{
exact( "string" )
}
\seealso{
\code{\link{extract}} \cr
  \code{\link[stringr]{fixed}} \cr
  \code{\link[stringr]{ignore.case}} \cr
  \code{\link[stringr]{perl}} \cr
}

