% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/searchable-package.R
\docType{package}
\name{searchable-package}
\alias{searchable-package}
\title{searchable}
\description{
Make R objects more searchable by allowing search of names based on case
(in)sensitivity, regular expressions, fixed patterns, exact matches, etc.
The package facilitates making and searching dictionaries,
}
\examples{
# Vectors
    v <- c( a=1, b=2, c=3 )
    sv <- searchable( v, ignore.case )
    sv[[ "b" ]]     # 2
    sv[[ ignore.case('B') ]]

  # LISTS
    l <- list( a=1, B=2, c=3 )
    sl <- searchable(l)
}
\references{
\url{http://stackoverflow.com/questions/5671719/case-insensitive-search-of-a-list-in-r} \cr
}
\seealso{
\code{\link{searchable}}
}

