% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/std.R
\name{std}
\alias{std}
\alias{std.SearchableOrPattern}
\alias{std.character}
\alias{std.default}
\title{Use/revert to standard matching}
\usage{
std(object, ...)

\method{std}{default}(object, ...)

\method{std}{character}(object, ...)

\method{std}{SearchableOrPattern}(object, ...)
}
\arguments{
\item{object}{to make specification}

\item{...}{additional arguments passes to \code{\link{pattern}}

\code{std}}
}
\description{
Creates or modifies the search type to use default R matching
}
\examples{
pat <- std("a")
  detect( c('alpha','beta'), pat )
}
\seealso{
\code{\link{pattern}}
}

